Updated Display's threads to a proper structure
All controller functionality is now contained in Display. Display now has a single game thread that plays the game based off of a program state variable. Full functionality for the game is complete but needs testing. Data-displaying tables are in Display, but need to be updated to display non-dummy data.